// packages/generator/src/emitter/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
  export type Document = ReadonlyDeep<{
    datamodel: Datamodel;
    schema: Schema;
    // mappings: Mappings;
    // additional props:
    modelMappings: ModelMapping[];
  }>;
  export type Enum = ReadonlyDeep<{
    name: string;
    // values: EnumValue[];
    dbName?: string | null;
    // documentation?: string;
    // additional props:
    typeName: string;
    docs: string | undefined;
    valuesMap: Array<{ name: string; value: string }>;
  }>;
// ...
}